Sample idea for the new metadata we will include
Purpose is to let us cheaply refer to developers, without requiring
their steamid for our auth system.

HOSTS/GUESTS will now support an extra {} field, which is a URL to an
avatar.

ADDITIONAL could probably be renamed, but it was always meant to be a
data field we would abuse. This should be a comma-delimited list of URLs
or Steam appid to fetch the game header.

YOUTUBE is new and will refer to the live video, so we can include a
link on the site.